consolidated sf types
