+++ apps/sim/tools/salesforce/delete_account.ts
@@ -1,26 +1,12 @@
 import { createLogger } from '@/lib/logs/console/logger'
+import type {
+  SalesforceDeleteAccountParams,
+  SalesforceDeleteAccountResponse,
+} from '@/tools/salesforce/types'
+import { getInstanceUrl } from '@/tools/salesforce/utils'
 import type { ToolConfig } from '@/tools/types'
-import { getInstanceUrl } from './utils'
 
 const logger = createLogger('SalesforceDeleteAccount')
-
-export interface SalesforceDeleteAccountParams {
-  accessToken: string
-  idToken?: string
-  instanceUrl?: string
-  accountId: string
-}
-
-export interface SalesforceDeleteAccountResponse {
-  success: boolean
-  output: {
-    id: string
-    deleted: boolean
-    metadata: {
-      operation: 'delete_account'
-    }
-  }
-}
 
 export const salesforceDeleteAccountTool: ToolConfig<
   SalesforceDeleteAccountParams,
