+++ apps/sim/tools/salesforce/delete_opportunity.ts
@@ -1,23 +1,9 @@
+import type {
+  SalesforceDeleteOpportunityParams,
+  SalesforceDeleteOpportunityResponse,
+} from '@/tools/salesforce/types'
+import { getInstanceUrl } from '@/tools/salesforce/utils'
 import type { ToolConfig } from '@/tools/types'
-import { getInstanceUrl } from './utils'
-
-export interface SalesforceDeleteOpportunityParams {
-  accessToken: string
-  idToken?: string
-  instanceUrl?: string
-  opportunityId: string
-}
-
-export interface SalesforceDeleteOpportunityResponse {
-  success: boolean
-  output: {
-    id: string
-    deleted: boolean
-    metadata: {
-      operation: 'delete_opportunity'
-    }
-  }
-}
 
 export const salesforceDeleteOpportunityTool: ToolConfig<
   SalesforceDeleteOpportunityParams,
