+++ apps/sim/tools/salesforce/update_contact.ts
@@ -1,37 +1,12 @@
 import { createLogger } from '@/lib/logs/console/logger'
+import type {
+  SalesforceUpdateContactParams,
+  SalesforceUpdateContactResponse,
+} from '@/tools/salesforce/types'
+import { getInstanceUrl } from '@/tools/salesforce/utils'
 import type { ToolConfig } from '@/tools/types'
-import { getInstanceUrl } from './utils'
 
 const logger = createLogger('SalesforceContacts')
-
-export interface SalesforceUpdateContactParams {
-  accessToken: string
-  idToken?: string
-  instanceUrl?: string
-  contactId: string
-  lastName?: string
-  firstName?: string
-  email?: string
-  phone?: string
-  accountId?: string
-  title?: string
-  department?: string
-  mailingStreet?: string
-  mailingCity?: string
-  mailingState?: string
-  mailingPostalCode?: string
-  mailingCountry?: string
-  description?: string
-}
-
-export interface SalesforceUpdateContactResponse {
-  success: boolean
-  output: {
-    id: string
-    updated: boolean
-    metadata: { operation: 'update_contact' }
-  }
-}
 
 export const salesforceUpdateContactTool: ToolConfig<
   SalesforceUpdateContactParams,
